\name{DMVI-package}
\alias{DMVI-package}
\alias{DMVI}
\docType{package}
\title{
\packageTitle{DMVI}
}
\description{
\packageDescription{DMVI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DMVI}
\packageIndices{DMVI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DMVI}

Maintainer: \packageMaintainer{DMVI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
